Remove duplicated IDs then remove link element IDs
Two of the HTML files had elements with duplicated ID attributes. The
thing with IDs is that they must be unique in a HTML document (so no
two elements can have the same IDs). The first thing I did was to remove
the duplicate.

Because IDs are typically used to uniquely identify an element for
scripting (JavaScript), CSS or as a fragment identifier I then removed
the IDs altogether because they didn't seem needed for these purposes.